strip upload row if unsuccessful
